+++ packages/components/frameworks/react-ssr/src/App.js
@@ -84,11 +84,16 @@ function App() {
       <div className="components-examples">
         {/* CURRENTLY TESTING */}
         <div className="example-wrapper">
-          {/* Test the component here. When done add it to the list alphabetically */}
+          {/* Test the component here (delete what was here previously). When done add it to the list alphabetically */}
           <h3>Test your component here</h3>
           {/* Normal version */}
           <div className="e-bg-white">
-            <Tabs items={tabsItems} value={2} valueOnChange={(event) => logValue('Tabs', event)}></Tabs>
+            <Tabs
+              items={tabsItems}
+              value={2}
+              valueOnChange={(event) => logValue('Tabs', event)}
+              hasManualActivation
+            ></Tabs>
           </div>
           {/* Inverted version */}
           <div className="e-bg-grey"></div>